removed ts treesitter, better windows experience

+++ setup.ps1
@@ -1,25 +1,27 @@
 param(
+  [switch]$vscode,
+  [switch]$theme,
   [switch]$wsl,
-  [switch]$neovim,
-  [switch]$theme
+  [switch]$neovim
 )
 
 # install vscode
-code --version >$null
-if (! $?){
-  $build = 'stable'
-  $os = 'win32-x64-user'
-  $SourceURL = "https://code.visualstudio.com/sha/download?build=$build&os=$os";
-  $Installer = $env:TEMP + "\vscode.exe";
-  $ProgressPreference = 'SilentlyContinue'
-  Invoke-WebRequest $SourceURL -OutFile $Installer;
-  Start-Process -FilePath $Installer -Args "/mergetasks=!runcode,addcontextmenufiles,addcontextmenufolders,addtopath,desktopicon" -Wait;
-  Remove-Item $Installer;
+if ($vscode){
+  code --version >$null
+  if (! $?){
+    $build = 'stable'
+    $os = 'win32-x64-user'
+    $SourceURL = "https://code.visualstudio.com/sha/download?build=$build&os=$os";
+    $Installer = $env:TEMP + "\vscode.exe";
+    $ProgressPreference = 'SilentlyContinue'
+    Invoke-WebRequest $SourceURL -OutFile $Installer;
+    Start-Process -FilePath $Installer -Args "/mergetasks=!runcode,addcontextmenufiles,addcontextmenufolders,addtopath,desktopicon" -Wait;
+    Remove-Item $Installer;
+  }
+  ./vscode/install_vscode_extensions.ps1
+  New-Item -ItemType HardLink -Path $env:APPDATA/Code/User/keybindings.json -Target $HOME/.terminal_setup/vscode/keybindings.json -Force
+  New-Item -ItemType HardLink -Path $env:APPDATA/Code/User/settings.json -Target $HOME/.terminal_setup/vscode/settings.json -Force
 }
-./vscode/install_vscode_extensions.ps1
-New-Item -ItemType HardLink -Path $env:APPDATA/Code/User/keybindings.json -Target $HOME/.terminal_setup/vscode/keybindings.json -Force
-New-Item -ItemType HardLink -Path $env:APPDATA/Code/User/settings.json -Target $HOME/.terminal_setup/vscode/settings.json -Force
-
 # set theme
 if ($theme){
 Copy-Item -Force $HOME/.terminal_setup/theme/DarkWolf.theme $HOME/AppData/Local/Microsoft/Windows/Themes/DarkWolf.theme
@@ -40,7 +42,12 @@ if ($neovim){
     return
   }
   winget install gnuwin32.make
+  # if some problem occurs while cloning neovim packages using git,
+  # try `git config --global http.sslbackend schannel`
+  # https://stackoverflow.com/questions/57327608/ssl-certificate-problem-self-signed-certificate-in-certificate-chain
   winget install Neovim.Neovim
+  # https://github.com/BurntSushi/ripgrep#installation
+  winget install BurntSushi.ripgrep.MSVC
   Remove-Item -Force -ErrorAction Ignore $env:LOCALAPPDATA/nvim
   New-Item -ItemType SymbolicLink -Path $env:LOCALAPPDATA/nvim -Target $HOME/.terminal_setup/linux_terminal/nvim
 }
